Move contract related functions to utils/chain.py
